added google login

+++ frontend/src/pages/LogOut.js
@@ -1,6 +1,13 @@
+import {useAuth0} from '@auth0/auth0-react';
+
 function LogOut() {
-  sessionStorage.clear();
-  window.location.href = "/";
+  const {logout} = useAuth0()
+  if (sessionStorage['email']) {
+    sessionStorage.clear();
+    window.location.href = "/";
+  } else {
+    logout();
+  }
 }
 
 export default LogOut
